+++ src/cpp/tardigrade_cmml_header_only_materials.h
@@ -1,0 +1,15 @@
+/*!
+=====================================================================
+|                tardigrade_header_only_materials.h                 |
+=====================================================================
+| Materials which are defined in header-only files. Allows for      |
+| materials to be registered in the library.                        |
+=====================================================================
+*/
+
+#ifndef TARDIGRADE_HEADER_ONLY_MATERIALS_H
+#define TARDIGRADE_HEADER_ONLY_MATERIALS_H
+
+    #include "tardigrade_BasicReactingSolid.h"
+
+#endif