+++ docs/sphinx/changelog.rst
@@ -11,6 +11,10 @@ Changelog
 
 New Features
 ============
+- Initial commit (:pull:`1`). By `Nathan Miller`_.
+- Added the basic reacting continuum class (:pull:`2`). By `Nathan Miller`_.
+- Added the full material model library build (:pull:`3`). By `Nathan Miller`_.
+- Added a function that returns the evaluate model results size (:pull:`4`). By `Nathan Miller`_.
 
 Internal Changes
 ================